pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentskills"
version = "1.0.0"
description = "Collection of Claude Code skills"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
research = ["httpx>=0.27.0"]
gamma = []  # only needs python-dotenv
keep = ["gkeepapi>=0.16.0", "gpsoauth>=1.0.0", "click>=8.0.0", "rich>=13.0.0"]
nanobanana = ["google-generativeai", "pillow"]
all = [
    "httpx>=0.27.0",
    "gkeepapi>=0.16.0",
    "gpsoauth>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "google-generativeai",
    "pillow",
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
research = "scripts.research:main"
poll_research = "scripts.poll_research:main"
extract_json = "scripts.extract_json:main"
generate_gamma_presentation = "scripts.generate_gamma_presentation:main"
get_gamma_assets = "scripts.get_gamma_assets:main"
google_keep = "scripts.google_keep:cli"
nanobanana = "scripts.nanobanana:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
